Move parsing of ifconfig output & etc/hosts for LocalHost to commonMain
