+++ library/runtime-ctrl-api/src/jsTest/kotlin/io/matthewnelson/kmp/tor/runtime/ctrl/api/address/LocalHostJsUnitTest.kt
@@ -15,34 +15,4 @@
  **/
 package io.matthewnelson.kmp.tor.runtime.ctrl.api.address
 
-import io.matthewnelson.kmp.tor.runtime.ctrl.api.internal.tryOsNetworkInterfaces
-import io.matthewnelson.kmp.tor.runtime.ctrl.api.internal.tryParseEtcHosts
-import io.matthewnelson.kmp.tor.runtime.ctrl.api.internal.tryParseIfConfig
-import kotlin.test.Test
-
-class LocalHostJsUnitTest: LocalHostBaseTest() {
-
-    @Test
-    fun givenUnixHost_whenOsNetworkInterfaces_thenReturnsLocalHostIPs() {
-        val set = LinkedHashSet<IPAddress>(2, 1.0F)
-        LocalHost.tryOsNetworkInterfaces(set)
-        // No assertions here as host machine running tests may not have
-        println(set)
-    }
-
-    @Test
-    fun givenUnixHost_whenIfConfig_thenReturnsLocalHostIPs() {
-        val set = LinkedHashSet<IPAddress>(2, 1.0F)
-        LocalHost.tryParseIfConfig(set)
-        // No assertions here as host machine running tests may not have
-        println(set)
-    }
-
-    @Test
-    fun givenUnixHost_whenEtcHosts_thenReturnsLocalHostIPs() {
-        val set = LinkedHashSet<IPAddress>(2, 1.0F)
-        LocalHost.tryParseEtcHosts(set)
-        // No assertions here as host machine running tests may not have
-        println(set)
-    }
-}
+class LocalHostJsUnitTest: LocalHostBaseTest()
