+++ library/runtime-ctrl-api/src/jvmMain/kotlin/io/matthewnelson/kmp/tor/runtime/ctrl/api/internal/-JvmPlatform.kt
@@ -19,10 +19,6 @@ import io.matthewnelson.kmp.file.SysPathSep
 import io.matthewnelson.kmp.tor.core.api.annotation.InternalKmpTorApi
 import io.matthewnelson.kmp.tor.core.resource.OSHost
 import io.matthewnelson.kmp.tor.core.resource.OSInfo
-import io.matthewnelson.kmp.tor.runtime.ctrl.api.address.IPAddress
-import io.matthewnelson.kmp.tor.runtime.ctrl.api.address.IPAddress.Companion.toIPAddress
-import io.matthewnelson.kmp.tor.runtime.ctrl.api.address.LocalHost
-import java.net.InetAddress
 
 @OptIn(InternalKmpTorApi::class)
 internal actual val UnixSocketsNotSupportedMessage: String? by lazy {
@@ -74,10 +70,4 @@ private val AndroidPID: Int? by lazy {
     } catch (_: Throwable) {
         null
     }
-}
-
-@Throws(Exception::class)
-internal actual fun LocalHost.Companion.resolveAllTo(set: LinkedHashSet<IPAddress>) {
-    val addresses = InetAddress.getAllByName("localhost")
-    addresses.mapTo(set) { it.hostAddress.toIPAddress() }
 }
